% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_get_sld_list}
\alias{db_get_sld_list}
\title{Get the list of what data is sideloaded for a patient}
\usage{
db_get_sld_list(runid)
}
\description{
Get the list of what data is sideloaded for a patient
}
