% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_get_variants}
\alias{db_get_variants}
\title{Get the variants with recur,gxp and CN from the database}
\usage{
db_get_variants(runid, tp)
}
\description{
Get the variants with recur,gxp and CN from the database
}
