% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_put_groupings}
\alias{db_put_groupings}
\title{Copy new groupings from a sheet to the database}
\usage{
db_put_groupings(sheet, db = Sys.getenv("CURVE_DB"))
}
\description{
Copy new groupings from a sheet to the database
}
