\name{effective.curvature}
\alias{effective.curvature}
\alias{curvature.stats}
\alias{R11T.c}
\alias{rel.curvature}

\title{Utilities Used by Curvature and Project}
\description{xxx}
\usage{
	effective.curvature(cur,QRofD, residual, s.sqr, show.details=F, warn=T)
	curvature.stats(cur, n, signif=0.05)
	R11T.c(R11.inv,p,pp)
	rel.curvature(s.sqr,R11,R2, show.extra.details=F, 
                     eps=sqrt(.Machine$double.eps))
}
\arguments {
	\item{x}{xxx}

}
\value{x}
\details{
	x
}
\seealso{
	\code{\link{curvature}}
	\code{\link{project}}
}
\keyword{DSE}
\keyword{ts}
