% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_cnv}
\alias{format_cnv}
\title{Format CNV Tables for display}
\usage{
format_cnv(d, dict, table = "segments")
}
\arguments{
\item{d}{cnv segments from the database}

\item{dict}{The curve data dictionary}
}
\value{
out$table cnv table for display

out$hide index of columns to hide
}
\description{
Format CNV Tables for display
}
