% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_fus}
\alias{format_fus}
\title{Format Fusion tables for display}
\usage{
format_fus(fus, dict, detail = FALSE)
}
\arguments{
\item{fus}{fusions from the database}

\item{dict}{The curve data dictionary}
}
\value{
out$table fus table for display

out$hide index of columns to hide
}
\description{
Format Fusion tables for display
}
