% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_gxp}
\alias{format_gxp}
\title{Format gene expression for display}
\usage{
format_gxp(gxp, dict, detail = FALSE)
}
\arguments{
\item{gxp}{Gene expression from vault or DB}

\item{dict}{The curve data dictionary}
}
\value{
out$table Gene expression table formatted for display

out$hide index of columns to hide
}
\description{
Format gene expression for display
}
