% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_sv}
\alias{format_sv}
\title{Format SV tables for display}
\usage{
format_sv(sv, dict, detail = FALSE)
}
\arguments{
\item{sv}{structural variants from the database}

\item{dict}{The curve data dictionary}
}
\value{
out$table sv table for display

out$hide index of columns to hide
}
\description{
Format SV tables for display
}
