% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_variants}
\alias{format_variants}
\title{Format variants for display}
\usage{
format_variants(vars, dict, tp, detail = FALSE)
}
\arguments{
\item{dict}{The curve data dictionary}

\item{detail}{Include columns from the show_detail column rather than the show_table}

\item{var}{Variants from vault or database}
}
\value{
out$table Variants formatted for display

out$hide index of columns to hide

out$renderpct index of columns to show as percentages
}
\description{
Format variants for display
}
