\name{hessian.TSestModel}
\alias{hessian.TSestModel}

\title{Calculate Hessian Matrix}
\description{Calculatate the hessian matrix of a function at a parameter value.}
\usage{
    \method{hessian}{TSestModel}(obj, ...)
}
\arguments {
    \item{x}{See genD.}
}
\value{xxx}
\details{
    This function uses genD.
}
\seealso{
    \code{\link{hessian}}
    \code{\link{genD}}
    \code{\link{span}}
}

%\keyword{DSE}
\keyword{ts}

