% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_cnv}
\alias{import_cnv}
\title{Import CNV segments from vault to database}
\usage{
import_cnv(vault, gid, type)
}
\arguments{
\item{vault}{a TPO vault object}

\item{gid}{the grouping id in the database}

\item{type}{somatic ('som') or germline ('ger') analysis}
}
\description{
Import CNV segments from vault to database
}
