% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_gxp}
\alias{import_gxp}
\title{Import gene expression from vault to database}
\usage{
import_gxp(vault, gid, type)
}
\arguments{
\item{vault}{a TPO vault object}

\item{gid}{the grouping id in the database}

\item{type}{'tumor' or 'normal'}
}
\description{
Import gene expression from vault to database
}
