% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_meta}
\alias{import_meta}
\title{Import grouping from vault metadata}
\usage{
import_meta(vault)
}
\arguments{
\item{vault}{a TPO vailt object}
}
\value{
id the "id" from the groupings table
}
\description{
Import grouping from vault metadata
}
