% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_structural}
\alias{import_structural}
\title{Import structural variants from a TPO vault}
\usage{
import_structural(vault, gid)
}
\arguments{
\item{vault}{a TPO vault object}

\item{gid}{The id from the groupings table for the analysis}
}
\description{
Import structural variants from a TPO vault
}
