% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_variants}
\alias{import_variants}
\title{Import variants from a TPO vault}
\usage{
import_variants(vault, tp, gid)
}
\arguments{
\item{vault}{a TPO vault object}

\item{tp}{type (somatic or germline)}

\item{gid}{The id from the groupings table for the analysis}
}
\description{
Import variants from a TPO vault
}
