% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{init_gxp_genes}
\alias{init_gxp_genes}
\title{Initialize the gxp_genes table using a GTF file}
\usage{
init_gxp_genes(gtf = "/tpo/refs/grch38/ensembl/grch38.97.clean.gtf")
}
\description{
Initialize the gxp_genes table using a GTF file
}
