\name{load.curvature.c}
\alias{load.curvature.c}
\alias{load.curvature.f}
\title{Load Curvature Code}
\description{Fortan or C code for curvature functions is loaded into S.}
\usage{
	load.curvature.c()
	load.curvature.f()
}
\sideffects{Fortan or C code for curvature functions is dynamically loaded.}
\examples{
#load.curvature.c() defunct
#load.curvature.f() defunct?
}
\keyword{DSE}
\keyword{ts}

