% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxp_module.R
\name{match_ntile}
\alias{match_ntile}
\title{Annotate a gxp dataset with the various ntile cutoffs}
\usage{
match_ntile(data, gxp_ntile)
}
\description{
Annotate a gxp dataset with the various ntile cutoffs
}
