\name{print.curvature}
\alias{print.curvature}
\alias{print.curvatureArray}
\title{Specific Methods for Print}
\description{See the generic function description.}
\usage{
    \method{print}{curvature}(x, ...)
    \method{print}{curvatureArray}(x, ...)
}
\arguments{
    \item{x}{a curvature object to be printed.}
    \item{...}{arguments to be passed to other methods.}
}
\seealso{ \code{\link{print}} }

%\keyword{DSE}
\keyword{ts}


