\name{curvature.TSestModel}
\alias{curvature.TSestModel}
\title{Curvature for a TSestModel}
\description{Calculate curvature for a TSestModel.}
\usage{
	curvature.TSestModel(emodel, warn=T, ...)  
}

\arguments {
	\item{emodel}{A TSestModel.}
	\item{warn}{Print additional warning messages.}
}

\details{
	None yet.
}
\examples{
	\dontrun{curvature(some model)}
}
\keyword{DSE}
\keyword{ts}

