\name{hessian}
\alias{hessian}
\alias{hessian.default}
\alias{hessian.TSestModel}

\title{Calculate Hessian Matrix}
\description{Calculatate the hessian matrix of a function at a parameter value.}
\usage{
	hessian(func, x, func.args=NULL, d=0.01, eps=1e-4,r=6)
}
\arguments {
	\item{x}{See genD.}
}
\value{xxx}
\details{
	This function uses genD.
}
\seealso{
	\code{\link{genD}}
	\code{\link{span}}
}

\keyword{DSE}
\keyword{ts}

